pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repoforgex"
version = "0.3.0"
description = "Automate creation, scaffold and sync many GitHub repos with AI-powered features (hardened, Docker, GitHub App)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1.7",
  "PyYAML>=6.0",
  "requests>=2.31.0",
  "flask>=2.2.5",
  "tenacity>=8.2.2",
  "PyJWT>=2.8.0",
  "pydantic>=1.10.9",
  "python-dotenv>=1.0.0",
]
[project.urls]
"Home" = "https://github.com/kabir308/repoforgex"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = [
    "E203", # whitespace before ':'
    "I001", # isort - handled by isort tool separately
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"config.py" = ["N805"]  # pydantic validators use cls

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
